+++ CHANGELOG.md
@@ -2,6 +2,8 @@ v0.2.2
 ======
 
 * Added Prim's Algorithm (`alg.prim`).
+* Added method to change a directed graph to an undirected graph
+  (`Digraph.asUndirected`).
 
 v0.2.1
 ======